refactor: :construction: Fixed TS errors.
