+++ src/scripts/api-observables.ts
@@ -1,0 +1,5 @@
+import { map, Observable } from "rxjs";
+
+export const flightDataObservable$ = new Observable((subscriber) =>
+    subscriber.next()
+);
