+++ src/scripts/api-service.ts
@@ -1,6 +1,6 @@
 import { flightData$ } from "./observables";
 import { mapObserver } from "./map-service";
-import { listObserver } from "./dom-manip";
+import { listObserver } from "./dom-manipulation";
 
 function showFlightDataAndPopulateMap(): void {
     flightData$.subscribe(mapObserver);
